Add test for iterable enum instance creation and move tests to test_enum.py
